fix: solve issues with delayed evaluation for context-related objects (#244)
* fix: solve issues with delayed evaluation for context-related objects
* refactor: remove resolver classes now fully replaced by Func<>
* test: improve code coverage for methods TryGetValue

+++ Expressif.Testing/Predicates/PredicationFactoryTest.cs
@@ -53,7 +53,7 @@ public class PredicationFactoryTest
     public void Instantiate_NumericEqualToVariableParameter_Valid()
     {
         var context = new Context();
-        context.Variables.Add<int>("myVar", 2);
+        context.Variables.Add<int>("myVar", 2m);
         var predicate = new PredicationFactory().Instantiate(new SinglePredication([new Function("EqualTo", new[] { new VariableParameter("myVar") })]), context);
         Assert.That(predicate, Is.Not.Null);
         Assert.Multiple(() =>
@@ -67,7 +67,7 @@ public class PredicationFactoryTest
     public void Instantiate_NumericEqualToObjectPropertyParameter_Valid()
     {
         var context = new Context();
-        context.CurrentObject.Set(new { Digits = 3 });
+        context.CurrentObject.Set(new { Digits = 3m });
         var predicate = new PredicationFactory().Instantiate(new SinglePredication([new Function("EqualTo", new[] { new ObjectPropertyParameter("Digits") })]), context);
         Assert.That(predicate, Is.Not.Null);
         Assert.Multiple(() =>
@@ -81,7 +81,7 @@ public class PredicationFactoryTest
     public void Instantiate_NumericEqualToObjectIndexParameter_Valid()
     {
         var context = new Context();
-        context.CurrentObject.Set(new List<int> { 0, 4 });
+        context.CurrentObject.Set(new List<decimal> { 0, 4 });
         var predicate = new PredicationFactory().Instantiate(new SinglePredication([new Function("EqualTo", new[] { new ObjectIndexParameter(1) })]), context);
         Assert.That(predicate, Is.Not.Null);
         Assert.Multiple(() =>
